% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataZentra.R
\name{getDataZentra}
\alias{getDataZentra}
\title{Function to get data from the API of zentracloud.com}
\usage{
getDataZentra(
  token,
  var = "Water Content",
  device_sn = "z6-10293",
  port,
  time_span = c(now(), now() - days(5))
)
}
\arguments{
\item{token}{token get from zentracloud API}

\item{var}{which variable you want to get they could be
"Air Temperature","Atmospheric Pressure","Battery Percent",
"Battery Voltage","Gust Speed","Lightning Activity",
"Lightning Distance","Logger Temperature","Max Precip Rate"
"Precipitation","RH Sensor Temp","Reference Pressure",
"Saturation Extract EC","Soil Temperature","Solar Radiation",
"VPD", "Vapor Pressure", "Water Content",
"Wind Direction","Wind Speed","X-axis Level"
"Y-axis Level"}

\item{device_sn}{the name of the z6 datalogger}

\item{port}{the port number of sensor}

\item{time_span}{vector of two having time range for which the data is required}
}
\value{

}
\description{
Function to get data from the API of zentracloud.com
}
\examples{
# Get data of VPD for the last five days
# token <- my_token
# VPD_data <- getDataZentra(token,var = "VPD",port =1)
}
