% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataZentra.R
\name{getDataZentra}
\alias{getDataZentra}
\title{Function to get data from the API of zentracloud.com}
\usage{
getDataZentra(
  token,
  var = "Water Content",
  port,
  time_span = c(now(), now() - days(5))
)
}
\arguments{
\item{token}{token get from zentracloud API}

\item{var}{which variable you want to get}

\item{port}{the port number of sensor}

\item{time_span}{vector of two having time range for which the data is required}
}
\value{

}
\description{
Function to get data from the API of zentracloud.com
}
\examples{
# Get data of VPD for the last five days
# token <- my_token
# VPD_data <- getDataZentra(token,var = "VPD",port =1)
}
